# apis/backend-api/settings/sdkwork-settings-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Settings Backend API",
    "version": "0.1.0",
    "description": "Backend API contract for SDKWork Settings tenant configuration, system settings, and configuration revision management. Authenticated with an administrator bearer token. All HTTP 2xx JSON responses use the SdkWorkApiResponse envelope ({ code, data, traceId }); all HTTP 4xx/5xx error responses use application/problem+json with a numeric ProblemDetail.code. See sdkwork-specs/API_SPEC.md."
  },
  "servers": [
    {
      "url": "https://api.sdkwork.com/sdkwork",
      "description": "SDKWork platform production endpoint"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/settings/v1/backend-api/tenant-configs": {
      "get": {
        "tags": ["settings"],
        "operationId": "tenantConfigs.list",
        "summary": "List tenant configurations",
        "description": "List tenant configurations. Supports optional namespace filtering and offset pagination.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "description": "Filter tenant configurations by namespace.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "One-based page index for offset pagination. Default 1.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Page size for offset pagination. Default 20, maximum 200.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant configuration list retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfigListResponse"
                },
                "example": {
                  "code": 0,
                  "data": {
                    "items": [
                      {
                        "id": 12884901888,
                        "tenantId": 1,
                        "namespace": "feature-flags",
                        "configKey": "beta-ui",
                        "configValue": { "enabled": true },
                        "valueType": "object",
                        "createdAt": "2026-07-01T02:30:00.000Z",
                        "updatedAt": "2026-07-01T02:30:00.000Z",
                        "createdBy": 1001,
                        "updatedBy": 1001
                      }
                    ],
                    "pageInfo": {
                      "mode": "offset",
                      "page": 1,
                      "pageSize": 20,
                      "totalItems": "1",
                      "totalPages": 1,
                      "hasMore": false
                    }
                  },
                  "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/settings/v1/backend-api/tenant-configs/{namespace}/{key}": {
      "get": {
        "tags": ["settings"],
        "operationId": "tenantConfigs.retrieve",
        "summary": "Retrieve a single tenant configuration",
        "description": "Retrieve a single tenant configuration identified by namespace and key.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/NamespacePath" },
          { "$ref": "#/components/parameters/KeyPath" }
        ],
        "responses": {
          "200": {
            "description": "Tenant configuration retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfigResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "put": {
        "tags": ["settings"],
        "operationId": "tenantConfigs.update",
        "summary": "Update a tenant configuration",
        "description": "Update a single tenant configuration identified by namespace and key.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/NamespacePath" },
          { "$ref": "#/components/parameters/KeyPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantConfigRequest"
              },
              "example": {
                "configValue": { "enabled": false },
                "valueType": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant configuration updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfigResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/settings/v1/backend-api/system-settings": {
      "get": {
        "tags": ["settings"],
        "operationId": "systemSettings.list",
        "summary": "List system settings",
        "description": "List system settings. Supports optional namespace and scope filtering and offset pagination.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "description": "Filter system settings by namespace.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "Filter system settings by scope.",
            "schema": {
              "type": "string",
              "enum": ["global", "region"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "One-based page index for offset pagination. Default 1.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Page size for offset pagination. Default 20, maximum 200.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System settings list retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingListResponse"
                },
                "example": {
                  "code": 0,
                  "data": {
                    "items": [
                      {
                        "id": 12884901888,
                        "namespace": "platform",
                        "settingKey": "maintenance-mode",
                        "settingValue": { "enabled": false },
                        "valueType": "object",
                        "scope": "global",
                        "scopeValue": null,
                        "createdAt": "2026-07-01T02:30:00.000Z",
                        "updatedAt": "2026-07-01T02:30:00.000Z",
                        "createdBy": 1001,
                        "updatedBy": 1001
                      }
                    ],
                    "pageInfo": {
                      "mode": "offset",
                      "page": 1,
                      "pageSize": 20,
                      "totalItems": "1",
                      "totalPages": 1,
                      "hasMore": false
                    }
                  },
                  "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/settings/v1/backend-api/system-settings/{namespace}/{key}": {
      "get": {
        "tags": ["settings"],
        "operationId": "systemSettings.retrieve",
        "summary": "Retrieve a single system setting",
        "description": "Retrieve a single system setting identified by namespace and key.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/NamespacePath" },
          { "$ref": "#/components/parameters/KeyPath" }
        ],
        "responses": {
          "200": {
            "description": "System setting retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "put": {
        "tags": ["settings"],
        "operationId": "systemSettings.update",
        "summary": "Update a system setting",
        "description": "Update a single system setting identified by namespace and key.",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          { "$ref": "#/components/parameters/NamespacePath" },
          { "$ref": "#/components/parameters/KeyPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSystemSettingRequest"
              },
              "example": {
                "settingValue": { "enabled": true },
                "valueType": "object",
                "scope": "region",
                "scopeValue": "cn-east-1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "System setting updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Administrator bearer token issued by SDKWork IAM. Required for all backend-api operations."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Validation error. ProblemDetail.code is 40001.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": {
              "type": "about:blank",
              "title": "Bad Request",
              "status": 400,
              "code": 40001,
              "detail": "Validation error",
              "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required. ProblemDetail.code is 40101.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": {
              "type": "about:blank",
              "title": "Unauthorized",
              "status": 401,
              "code": 40101,
              "detail": "Authentication required",
              "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Permission denied. ProblemDetail.code is 40301.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": {
              "type": "about:blank",
              "title": "Forbidden",
              "status": 403,
              "code": 40301,
              "detail": "Permission required",
              "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error. ProblemDetail.code is 50001.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" },
            "example": {
              "type": "about:blank",
              "title": "Internal Server Error",
              "status": 500,
              "code": 50001,
              "detail": "Internal error",
              "traceId": "0195f2a0-7c44-7b2e-9f3a-2a6f5d8e91ab"
            }
          }
        }
      }
    },
    "schemas": {
...
      "TenantConfigListResponse": {
        "description": "Tenant configuration list response envelope.",
        "allOf": [
          { "$ref": "#/components/schemas/SdkWorkApiResponse" },
          {
            "type": "object",
            "required": ["data"],
            "properties": {
              "data": {
                "allOf": [
                  { "$ref": "#/components/schemas/SdkWorkPageData" },
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "items": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/TenantConfig"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
...
    }
  }
}